Implement annotations, pretty much according to the original design notes.
You inherit from grok.Annotation. The resulting class can be thought of as a
persistent adapter. It gets instantiated the first time an object is adapted
and then persisted as part of the object.  Like with adapters, use grok.context,
grok.implements/grok.provides to specify what's adapted and what's provided.
Use grok.name to specify the annotation key, which is always a good idea if
you anticipate moving the class around, because otherwise the annotation key
will be the dotted class path.