grok2html has been altered to use urllib to access Jim's buildout tutorial from svn.zope.org.
Dependency on Pygments has been added to doc/setup.py and the file
pygments_code_block_directive.py has been borrowed from
http://docutils.sourceforge.net/sandbox/code-block-directive in order to handle
the code-block directive used in Jim's buildout tutorial.

This still doesn't support highlighting because Jim uses::

    code-block: Python

and pygments expects::

    code-block: python  # (lower-case P)

Furthermore to have highlighting additional css directives are necessary to
colourise the code block. I haven't implemented this either.

I'm happy with the hack as it is because we may hope and expect code
highlighting to be included in stock docutils in the future.